+++ draft-packages/button/KaizenDraft/Button/IconButton.tsx
@@ -1,6 +1,9 @@
 import * as React from "react"
 import GenericButton, { IconButtonProps } from "./components/GenericButton"
 
+/**
+ * @deprecated draft-packages IconButton is deprecated. Please use IconButton from "@kaizen/button" instead.
+ */
 const IconButton: React.FunctionComponent<IconButtonProps> = (
   props: IconButtonProps
 ) => <GenericButton iconButton {...props} />