+++ draft-packages/menu/KaizenDraft/Menu/MenuDropdown.tsx
@@ -106,6 +106,9 @@ const MenuDropdown = ({
   }, [autoHide, handleDocumentClickForAutoHide])
 
   return (
+    // Disabling these because we don't want this to be keyboard focusable.
+    // Esc keypress should be used instead for the same behaviour (hasn't been implemented here yet)
+    // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
     <div
       id={id}
       ref={setPopperElement}