+++ draft-packages/tooltip/docs/Tooltip.stories.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable jsx-a11y/anchor-is-valid */
 import * as React from "react"
 import informationIcon from "@kaizen/component-library/icons/information-white.icon.svg"
 import meatballsIcon from "@kaizen/component-library/icons/meatballs.icon.svg"
@@ -23,7 +24,7 @@ export default {
      * To cater for false positives when the tooltip renders
      * with a different alignment (controlled by react-popper).
      */
-    chromatic: { diffThreshold: 0.7 },
+    chromatic: { diffThreshold: 1 },
     docs: {
       description: {
         component: 'import { Tooltip } from "@kaizen/draft-tooltip"',