+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.0...@kaizen/draft-modal@8.4.1) (2022-02-15)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+# [8.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.3.2...@kaizen/draft-modal@8.4.0) (2022-02-15)
+
+
+### Features
+
+* Install eslint-plugin-jsx-a11y ([#2527](https://github.com/cultureamp/kaizen-design-system/issues/2527)) ([f106089](https://github.com/cultureamp/kaizen-design-system/commit/f1060891a8771d61ec6329f9e854e15683b30382))
+
+
+
+
+
 ## [8.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.3.1...@kaizen/draft-modal@8.3.2) (2022-02-14)
 
 **Note:** Version bump only for package @kaizen/draft-modal