+++ draft-packages/table/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.12.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.11.2...@kaizen/draft-table@3.12.0) (2022-02-15)
+
+
+### Features
+
+* Install eslint-plugin-jsx-a11y ([#2527](https://github.com/cultureamp/kaizen-design-system/issues/2527)) ([f106089](https://github.com/cultureamp/kaizen-design-system/commit/f1060891a8771d61ec6329f9e854e15683b30382))
+
+
+
+
+
+## [3.11.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.11.1...@kaizen/draft-table@3.11.2) (2022-02-15)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
 ## [3.11.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.11.0...@kaizen/draft-table@3.11.1) (2022-02-14)
 
 **Note:** Version bump only for package @kaizen/draft-table