+++ draft-packages/tile/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.1...@kaizen/draft-tile@5.6.2) (2022-02-15)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [5.6.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.0...@kaizen/draft-tile@5.6.1) (2022-02-15)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
 # [5.6.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.5.4...@kaizen/draft-tile@5.6.0) (2022-02-10)