+++ packages/button/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/button@0.1.0...@kaizen/button@0.2.0) (2022-02-15)
+
+
+### Features
+
+* Install eslint-plugin-jsx-a11y ([#2527](https://github.com/cultureamp/kaizen-design-system/issues/2527)) ([f106089](https://github.com/cultureamp/kaizen-design-system/commit/f1060891a8771d61ec6329f9e854e15683b30382))
+
+
+
+
+
 # [0.1.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/button@0.0.1...@kaizen/button@0.1.0) (2022-02-10)