+++ packages/tabs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.3.2...@kaizen/tabs@1.3.3) (2022-02-15)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.3.1...@kaizen/tabs@1.3.2) (2022-02-15)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
 ## [1.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.3.0...@kaizen/tabs@1.3.1) (2022-02-11)
 
 **Note:** Version bump only for package @kaizen/tabs